contributors/adobe/incoming/css3-regions/interactivity/mouse/regions-mouse-events-006.html

<!DOCTYPE HTML>
<html>
    <head>
        <title>CSS Regions: ondblclick events work inside region and outside region</title>
        <link rel="author" title="David Alcala" href="mailto:dalcala@adobe.com">
        <link rel="help" href="http://www.w3.org/TR/css3-regions/">
        <link rel="help" href="http://www.w3.org/TR/css3-regions/#the-flow-into-property">
        <link rel="help" href="http://www.w3.org/TR/css3-regions/#the-flow-from-property">
        <link rel="help" href="http://www.w3.org/TR/css3-regions/#relation-to-document-events">
        <meta name="assert" content="CSS regions module does not alter the normal processing of events in the document
         tree. Therefore, mouse ondblclick events inside and outside a named flow should function correctly.">
        <meta name="flags" content="dom interact">
        <link href="support/css/regions-mouse-events.css" rel="stylesheet" type="text/css">
        <link href="/resources/testharness.css" rel="stylesheet" type="text/css">
        <script src="/resources/testharness.js"></script>
    </head>
    <body>
        <p id="msg">
            Double-click the mouse on the black square, then double-click the mouse on the orange square.
        </p>
        <div id="content">
            <div class="block" title="FAIL"></div>
            <div class="block" title="FAIL"></div>
            <div id="target-block" class="block"></div>
        </div>
        <div class="region"></div>
        <div class="region"></div>
        <div class="region"></div>
        <br/>
        <div id="outside-region">This div is outside the region</div>
        <div id="log"></div>
    </body>
    <script type="text/javascript">
        
        var testTimeout = 10000;
        setup({timeout: testTimeout});
        
        // This block is executed if running in WebKit's harness
        if (window.testRunner)
        {
        	testRunner.dumpAsText(false);
        }

        // Verify that CSS Regions are enabled in the browser.
        // Divs will be horizontal if Regions are enabled.
        // Divs will be vertical if Regions are not enabled.
        function checkLeftPosition(elemID) 
        {
            return document.getElementById(elemID).getBoundingClientRect().left;
        }
        
        if(checkLeftPosition("target-block") != 0) {

            var test1 = async_test("ondblclick event fired inside region", {timeout: testTimeout});
                test1.step( function() {
                    var testTarget = document.getElementById("target-block");
                    testTarget.ondblclick = test1.step_func( function(evt) {
                        /* Log assertion that the event got fired */
                        assert_true(true);
                        
                        /* Verify the target inside the region has the correct ID */
                        test(function() { assert_equals(evt.target.id, "target-block")}, "Target double-clicked on inside region has the correct ID");
                        
                        test1.done();
                    });
                    
                    mouseDblClick("target-block");
                });
                
            var test2 = async_test("ondblclick event fired outside region", {timeout: testTimeout});
                test2.step( function() {
                    var testTarget = document.getElementById("outside-region");
                    testTarget.ondblclick = test2.step_func( function(evt) {
                        /* Log assertion that the event got fired */
                        assert_true(true);
                        
                        /* Verify the target outside the region has the correct ID  */
                        test(function() { assert_equals(evt.target.id, "outside-region")}, "Div double-clicked on outside region has the correct ID");
                        
                        test2.done();

                        //Update the message stating that tests are complete
                        var msg = document.getElementById("msg");

                        var complete = document.createElement("p");
                        complete.innerHTML = "Tests are complete. All results in the Details section below should PASS.";
                        complete.style.color = "blue";
                        msg.appendChild(complete);

                    });
                    
                    mouseDblClick("outside-region");
                });

            function mouseDblClick(block)
            {
            	if(window.testRunner)
               	{
            		var startNode = document.getElementById(block);
            		var xStartPosition = startNode.getBoundingClientRect().left;
            		var yStartPosition = startNode.getBoundingClientRect().top;
            		eventSender.mouseMoveTo(xStartPosition, yStartPosition);
            		eventSender.mouseDown();
                    eventSender.mouseUp();
                    eventSender.mouseDown();
                    eventSender.mouseUp();
            	}
            }

        }
        else { 
            test(function() {assert_true(false) }, "Regions are not enabled");
        }

    </script>
</html>
